# Aggregate IndexingMsg