Switch from unorderedCopy (or plain assignment) to aggregated copies.
This has significant performance benefits, particularly on non-cray
systems.

Here's performance for the gather/scatter benchmarks on 16 nodes of an
FDR InfiniBand cluster:

Run with `--trials 1 --index-size 2000000`:

| benchmark  | before     | after      |
| ---------- | ---------- | ---------- |
| gather.py  | 0.03 GiB/s | 1.71 GiB/s |
| scatter.py | 0.04 GiB/s | 2.64 GiB/s |

And run with default problem size of `100,000,000` ('before' timed out
after 15 minutes):

| benchmark  | before     | after      |
| ---------- | ---------- | ---------- |
| gather.py  | N/A        | 3.21 GiB/s |
| scatter.py | N/A        | 4.86 GiB/s |

This also results in a ~2x speedup for scatter on a Cray-XC at 16 nodes.
There's no change for gather, since newSrcAggregator is still using
unorderedCopy for ugni.

Resolves https://github.com/mhmerrill/arkouda/issues/210